Add the 'img' folder to the 'wwwroot'.
Add the 'img' folder to the 'wwwroot', this folder is provided in the Microsoft guided project and can be found in the github repository (https://github.com/MicrosoftDocs/mslearn-interact-with-data-blazor-web-apps).
This folder contains image to use with the list of PizzaSpecial defined in the 'Home.razor' file. Also contains some icons to be used during the project.